+++ src/main/java/com/example/dev_web_gerador/Controller/EpicoController.java
@@ -1,8 +1,11 @@
 package com.example.dev_web_gerador.Controller;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import com.example.dev_web_gerador.Model.*;
+import com.example.dev_web_gerador.Repository.*;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,12 +21,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.dev_web_gerador.DTO.EpicoDTO;
 import com.example.dev_web_gerador.DTO.EpicoInputDTO;
-import com.example.dev_web_gerador.Model.Epico;
-import com.example.dev_web_gerador.Model.Projeto;
-import com.example.dev_web_gerador.Model.TipoEpico;
-import com.example.dev_web_gerador.Repository.EpicoRepository;
-import com.example.dev_web_gerador.Repository.ProjetoRepository;
-import com.example.dev_web_gerador.Repository.TipoEpicoRepository;
 import com.example.dev_web_gerador.codes.StatusCode;
 import com.example.dev_web_gerador.lib.ArvoreBinariaExemplo;
 
@@ -39,6 +36,12 @@ public class EpicoController {
 
     @Autowired
     ProjetoRepository projetoRepository;
+
+    @Autowired
+    HistoriaUsuarioRepository historiaUsuarioRepository;
+
+    @Autowired
+    TipoHistoriaUsuarioRepository tipoHistoriaUsuarioRepository;
 
     @Autowired
     ArvoreBinariaExemplo<EpicoDTO> arvoreBinaria;
@@ -66,6 +69,8 @@ public class EpicoController {
 
         Epico epicoSalvo = epicoRepository.save(epico);
 
+        gerarHistoriaUsuario(epicoSalvo.getId());
+
         Epico epicoArvore = epicoRepository.findById(epicoSalvo.getId()).get();
         EpicoDTO epicoc = new EpicoDTO();
         epicoc.setId(epicoArvore.getId());
@@ -87,6 +92,44 @@ public class EpicoController {
         return ResponseEntity.status(HttpStatus.CREATED).body("Épico adicionado à árvore com sucesso.");
     }
 
+
+    public ResponseEntity<List<HistoriaUsuario>> gerarHistoriaUsuario(@PathVariable(value = "epico_id") long epico_id) {
+        Epico epico = epicoRepository.findById(epico_id).get(); //Prourando o epico pelo id passado
+        Long tipoEpicoId = epico.getTipoEpico().getId(); //Armazenando o id
+
+
+        List<TipoHistoriaUsuario> tipoHistoriaUsuario = tipoHistoriaUsuarioRepository.findByTipoEpicoId(tipoEpicoId);
+
+        String epicoDescricao = epico.getDescricao(); //Armazenando a desccrição do épico para ser modificada e adicionada no hu
+        List<HistoriaUsuario> historias = new ArrayList<HistoriaUsuario>();
+        tipoHistoriaUsuario.forEach(tipo -> { //Pegando ccada historia retornada na lista
+            String palavra = epicoDescricao.replaceAll("(?<=\\bdesejo\\s)\\w+", tipo.getDescricao()); //Substituindo a palavra desejo pelo verbo
+
+            HistoriaUsuario historiaUsuario = salvarHistoriaUsuario(epico.getId(), epico, palavra, tipo.getId());
+
+            historias.add(historiaUsuario);
+
+        });
+
+        return ResponseEntity.ok(historias);
+    }
+
+
+    private HistoriaUsuario salvarHistoriaUsuario(long epico_id,Epico epico, String descricao, long tipoHistoriaUsuario) {
+        HistoriaUsuario historiaUsuario = new HistoriaUsuario();
+
+        Optional<Epico> epicoOptional = epicoRepository.findById(epico.getId());
+        Optional<TipoHistoriaUsuario> tipoHistoriaUsuarioOptional = tipoHistoriaUsuarioRepository.findById(tipoHistoriaUsuario);
+        historiaUsuario.setCategoria(epico.getCategoria());
+        historiaUsuario.setDescricao(descricao);
+        historiaUsuario.setRelevancia(epico.getRelevancia());
+        historiaUsuario.setTitulo(epico.getTitulo());
+        historiaUsuario.setEpico(epicoOptional.get());
+        historiaUsuario.setTipoHistoriaUsuario(tipoHistoriaUsuarioOptional.get());
+
+        return historiaUsuarioRepository.save(historiaUsuario);
+    }
+
     @DeleteMapping("/removerDaArvore/{id}")
     public ResponseEntity<Object> removerEpicoDaArvore(@PathVariable long id) { // Remover o epico da arvore
         EpicoDTO epicoRemovido = arvoreBinaria.removerPorId(id);
@@ -105,13 +148,11 @@ public class EpicoController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Object> epicoId(@PathVariable long id) {
-        EpicoDTO epicoEncontrado = arvoreBinaria.pesquisarPorId(id);
+        Optional<Epico> epico = epicoRepository.findById(id);
 
-        if (epicoEncontrado != null) {
-            return ResponseEntity.status(HttpStatus.OK).body(epicoEncontrado);
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(StatusCode.EPIC_NOT_FOUND.getCode());
-        }
+        return epico.<ResponseEntity<Object>>map(epicos ->
+                ResponseEntity.status(HttpStatus.OK).body(epicos)).orElseGet(() ->
+                ResponseEntity.status(HttpStatus.NOT_FOUND).body(StatusCode.EPIC_NOT_FOUND.getCode()));
     }
 
     @PutMapping("/{id}")
